requirements.txt
fastapi 
uvicorn 
psycopg2 
bcrypt
python-dotenv
passlib
email-validator
starlette
passlib[argon2]